Extract signature verification & pass sigchecker to stackmanager
