+++ divi/src/script/script_error.h
@@ -54,4 +54,18 @@ typedef enum ScriptError_t
 
 const char* ScriptErrorString(const ScriptError error);
 
+static inline bool set_success(ScriptError* ret)
+{
+    if (ret)
+        *ret = SCRIPT_ERR_OK;
+    return true;
+}
+
+static inline bool set_error(ScriptError* ret, const ScriptError serror)
+{
+    if (ret)
+        *ret = serror;
+    return false;
+}
+
 #endif // BITCOIN_SCRIPT_SCRIPT_ERROR_H
